Changed Download function to Portal class
