+++ src/main/java/uk/gov/hmcts/darts/automation/pageObjects/Portal.java
@@ -2,15 +2,19 @@ package uk.gov.hmcts.darts.automation.pageObjects;
 
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Table;
+import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -134,5 +138,34 @@ public class Portal {
     public void notificationCount(String count) {
         NAV.waitForPageLoad();
         webDriver.findElement(By.xpath("//span[contains(@id, 'notifications') and contains(text(),'"+count+"')]"));
+    }
+
+    public void downloadFileMatches(String fileName) throws IOException {
+        String workspace_dir = ReadProperties.getDownloadFilepath();
+        File directory = new File(workspace_dir);
+        boolean downloadinFilePresence = false;
+        File[] filesList =directory.listFiles();
+        if(Objects.nonNull(filesList)) {
+            for (File file : filesList) {
+                downloadinFilePresence = file.getName().equalsIgnoreCase(fileName);
+                if (downloadinFilePresence) {
+                    log.info("File downloaded {} found and matched as expected", fileName);
+                    break;
+                } else {
+                    log.error("File {} is not downloaded and cannot be found", fileName);
+                }
+            }
+        }
+        Assert.assertTrue(downloadinFilePresence);
+        deleteDocument_withName_fromDownloads(workspace_dir);
+    }
+    private void deleteDocument_withName_fromDownloads(String workspace_dir) throws IOException {
+        try {
+            FileUtils.cleanDirectory(new File(workspace_dir));
+            log.info("Successfully deleted all documents from the directory: {}", workspace_dir);
+        } catch (IOException e) {
+            log.error("Error occurred while cleaning the directory: {}", workspace_dir, e);
+            throw e;
+        }
     }
 }
